% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.ICD10.R
\name{build_concept_tables.NHSICD10}
\alias{build_concept_tables.NHSICD10}
\title{Build ICD10 concept tables (see build_concept_tables)}
\usage{
build_concept_tables.NHSICD10(dict, replacements)
}
\description{
Build ICD10 concept tables (see build_concept_tables)
}
