% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_concept_tables}
\alias{build_concept_tables}
\title{Builds chosen dictionary concept tables in a database. Uses the supplied dictionary object to
determine how to do this. The concept_file_path specifies the location of the downloaded clinical code files.
<data-file-path> placeholders in the SQL file are replaced with the supplied concept_file_path.}
\usage{
build_concept_tables(dict, concept_file_path)
}
\arguments{
\item{dict}{Dictionary object}

\item{concept_file_path}{Path to the downloaded concept files}
}
\description{
Internally the function executes specific SQL files before and after a dictionary/database specific function (if it exists).
The form of this function is build_concept_tables.db_type.dict_type. It is passed the dictionary object and the list of replacements.
}
