% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{cc_debug}
\alias{cc_debug}
\title{Utility function which switches on/off debugging messages and returns the current status of debugging}
\usage{
cc_debug(value = NULL)
}
\arguments{
\item{value}{Boolean (T/F or TRUE/FALSE)}
}
\description{
Utility function which switches on/off debugging messages and returns the current status of debugging
}
\examples{

#switch debugging on
cc_debug(T)

#log something if debugging is on
if(cc_debug()) {
  print("debugging is on")
}

}
