% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{cc_disconnect}
\alias{cc_disconnect}
\title{Utility function used to disconnect database connection}
\usage{
cc_disconnect(dict)
}
\arguments{
\item{dict}{Clinical dictionary object}
}
\description{
Utility function used to disconnect database connection
}
