% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{cc_from_home}
\alias{cc_from_home}
\title{This function will use a JSON file in the users home directory to build a clinical dictionary object. This is used by subsequent functions
to direct building and querying.
The location of the JSON file will depend on the operating system being used.
On Linux/OS X this will be /Users/<user>/. On newer versions of Windows it should be the same, although it may change.
To determine where the file will need to be placed in R you can execute: \code{Sys.getenv('USERPROFILE')}}
\usage{
cc_from_home(dict_type, filename)
}
\arguments{
\item{dict_type}{Clinical concept dictionary type. Currently supported are NHSReadV2, NHSReadV3, NHSICD10 and NHSSnomedCT}

\item{filename}{Name of JSON file in user's home directory containing options to use to create a clinical dictionary object}
}
\description{
For security reasons, this function and \code{\link{cc_from_file}} are preferred to \code{\link{cc_from_list}}, as this prevents usernames and passwords being included in source code.
}
\examples{
\dontrun{
dict<-cc_from_home("NHSReadV3","dictconfig.ini")

#clinical dictionary object then passed to other functions
...
}
}
\seealso{
\code{\link{cc_from_list}}, \code{\link{cc_from_file}}
}
