% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{cc_from_list}
\alias{cc_from_list}
\title{Create a clinical dictionary object using a list of options. This is used by subsequent functions
to direct building and querying.}
\usage{
cc_from_list(dict_type, options)
}
\arguments{
\item{dict_type}{Clinical concept dictionary type. Currently supported are NHSReadV2, NHSReadV3, NHSICD10 and NHSSnomedCT}

\item{options}{List of options to use to create a clinical dictionary object

Global options (* specifies required options)
  type - specifies the database management system being used. Currently supported are mysql and sqlite [default])
  dbname* - when mysql this specifies the database name; when sqlite specifies the location of the database file

  user - database user
  pass - database password
  host - database host (default localhost)
  port - database port (default 3306)}
}
\description{
The preferred approach is to use \code{\link{cc_from_file}} as this is more secure.
}
\examples{
dict<-cc_from_list(list(type="sqlite",dbname="/path/to/sqlitedb.sqlite"))

#clinical dictionary object then passed to other functions
...


}
\seealso{
\code{\link{cc_from_file}}, \code{\link{cc_from_home}}
}
