% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{cc_is_available}
\alias{cc_is_available}
\title{Utility function to determine if clinical dictionary connection is available/connects}
\usage{
cc_is_available(dict, output_error = F)
}
\arguments{
\item{dict}{Clinical dictionary object}

\item{output_error}{When TRUE will output error message if is not available}
}
\value{
TRUE if connection is available
}
\description{
Utility function to determine if clinical dictionary connection is available/connects
}
