% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{cc_reconnect}
\alias{cc_reconnect}
\title{Utility function to reconnect if clinical dictionary connection has timed out}
\usage{
cc_reconnect(dict)
}
\arguments{
\item{dict}{Clinical dictionary object}
}
\description{
Utility function to reconnect if clinical dictionary connection has timed out
}
