% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{create_query_params_for_children_from_hierarchy}
\alias{create_query_params_for_children_from_hierarchy}
\title{Internal function used to create query parameters for getting children from a hierachical dictionary such as READ v2 or ICD10.}
\usage{
create_query_params_for_children_from_hierarchy(code, immediate = F)
}
\arguments{
\item{code}{Code to use as a basis for query parameters}

\item{immediate}{T/F flag to return just immediate children or all children}
}
\value{
List containing dplyr::filter query op and value
}
\description{
Internal function used to create query parameters for getting children from a hierachical dictionary such as READ v2 or ICD10.
}
