% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{disable_case_sensitivity}
\alias{disable_case_sensitivity}
\title{Disable SQLite database query case-sensitivity}
\usage{
disable_case_sensitivity(dict)
}
\arguments{
\item{dict}{dictionary object (see cc_from_list/cc_from_file)}
}
\description{
Disable SQLite database query case-sensitivity
}
