% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{enable_case_sensitivity}
\alias{enable_case_sensitivity}
\title{Enable SQLite database query case-sensitivity}
\usage{
enable_case_sensitivity(dict)
}
\arguments{
\item{dict}{dictionary object (see cc_from_list/cc_from_file)}
}
\description{
Enable SQLite database query case-sensitivity
}
