% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{execute_sql_statements}
\alias{execute_sql_statements}
\title{Utility function which loops through a vector of SQL statements and executes them}
\usage{
execute_sql_statements(dict, statements)
}
\arguments{
\item{dict}{Dictionary object}

\item{statements}{Vector of SQL statements}
}
\description{
Utility function which loops through a vector of SQL statements and executes them
}
