% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{extract_relations_from_dag}
\alias{extract_relations_from_dag}
\title{Internal function used to get relations from a diacyclic graph-type dictionary such as READ v3 or SNOMED-CT.}
\usage{
extract_relations_from_dag(dict, code, immediate_relations, children)
}
\arguments{
\item{dict}{Clinical dictionary object}

\item{code}{Code to use a basis for relationship extraction}

\item{immediate_relations}{T/F flag to return just immediate relations or all relations}

\item{children}{T/F flag to return children (T) or parents (F)}
}
\value{

}
\description{
Internal function used to get relations from a diacyclic graph-type dictionary such as READ v3 or SNOMED-CT.
}
