% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{extract_relations_from_hierarchy}
\alias{extract_relations_from_hierarchy}
\title{Internal function used to get relations from a hierarchical dictionary such as READ v2 or ICD10}
\usage{
extract_relations_from_hierarchy(dict, code, immediate_relations = F,
  children = T)
}
\arguments{
\item{dict}{Clinical dictionary object}

\item{code}{Code to use a basis for relationship extraction}

\item{immediate_relations}{T/F flag to return just immediate relations or all relations}

\item{children}{T/F flag to return children (T) or parents (F)}
}
\description{
Internal function used to get relations from a hierarchical dictionary such as READ v2 or ICD10
}
