% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{find_matching_file}
\alias{find_matching_file}
\title{Find file matching regular expression in directory. Used to retrieve code files which vary depending on version.}
\usage{
find_matching_file(file_path, match)
}
\arguments{
\item{file_path}{Path to search for file}

\item{match}{Regular expression to use}
}
\value{
File name
}
\description{
Find file matching regular expression in directory. Used to retrieve code files which vary depending on version.
}
\examples{
find_matching_file("/path/to/data","/description/")

}
