% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.ICD10.R
\name{get_child_codes.NHSICD10}
\alias{get_child_codes.NHSICD10}
\title{Get ICD10 child codes (see get_child_codes)}
\usage{
\method{get_child_codes}{NHSICD10}(dict, code, immediate_children = F,
  current_only = F)
}
\description{
Get ICD10 child codes (see get_child_codes)
}
