% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child_codes.R
\name{get_child_codes.NHSReadV3}
\alias{get_child_codes.NHSReadV3}
\title{Internal function user to get descendent read v2 codes from the database (called explicitly by get_child_readcodes)}
\usage{
\method{get_child_codes}{NHSReadV3}(dict, code, immediate_children = F,
  current_only = F)
}
\arguments{
\item{immediate_children}{indicates if only one level (e.g. just children) should be returned}

\item{config}{rexceed config object (see new_config)}

\item{read_code}{READ code to obtain descendents for}
}
\value{
character vector including descendent codes (excluding read_code provided)
}
\description{
Internal function user to get descendent read v2 codes from the database (called explicitly by get_child_readcodes)
}
\examples{
config<-rexceed::rexceed_config_from_home("20160103")
h3_v2_children<-rexceed::ref_get_child_v2_readcodes(config,"H3...",TRUE)
}
