% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{get_child_codes}
\alias{get_child_codes}
\title{Get child codes from the supplied clinical dictionary}
\usage{
get_child_codes(dict, code, immediate_children = F, active_only = T)
}
\arguments{
\item{dict}{Clinical dictionary object}

\item{code}{Clinical code to get child codes for}

\item{immediate_children}{T/F flag to return only children from the first level below (default FALSE)}

\item{active_only}{T/F flag to return only active codes (default TRUE)}
}
\description{
Get child codes from the supplied clinical dictionary
}
\examples{
\dontrun{
dict<-cc_from_file("/path/to/dictconfig")
h3_child_codes<-get_parent_codes(dict,"H31..",immediate_children=F)
}
}
