% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{get_ctable_code_field}
\alias{get_ctable_code_field}
\title{Get dictionary-specific concept code field name}
\usage{
get_ctable_code_field(dict)
}
\arguments{
\item{dict}{Clinical dictionary object}
}
\value{
code field name
}
\description{
Get dictionary-specific concept code field name
}
