% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{get_ctable_name}
\alias{get_ctable_name}
\title{Get dictionary-specific concept table name}
\usage{
get_ctable_name(dict)
}
\arguments{
\item{dict}{Clinical dictionary object}
}
\value{
table name
}
\description{
Get dictionary-specific concept table name
}
