% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.ICD10.R
\name{get_ctable_term_field.NHSICD10}
\alias{get_ctable_term_field.NHSICD10}
\title{Get concept code term name in ICD10 concept table}
\usage{
\method{get_ctable_term_field}{NHSICD10}(dict)
}
\description{
Get concept code term name in ICD10 concept table
}
