% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{get_ctable_term_field}
\alias{get_ctable_term_field}
\title{Get dictionary-specific concept term field name}
\usage{
get_ctable_term_field(dict)
}
\arguments{
\item{dict}{Clinical dictionary object}
}
\value{
term field name
}
\description{
Get dictionary-specific concept term field name
}
