% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{get_dot_position}
\alias{get_dot_position}
\title{Internal function which gets the position of first dot in code}
\usage{
get_dot_position(code)
}
\arguments{
\item{code}{Code to search}
}
\value{
An integer specifying the location of the first dot
}
\description{
Internal function which gets the position of first dot in code
}
