% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.ICD10.R
\name{get_parent_codes.NHSICD10}
\alias{get_parent_codes.NHSICD10}
\title{Get ICD10 parent codes (see get_parent_codes)}
\usage{
\method{get_parent_codes}{NHSICD10}(dict, code, immediate_parents = F,
  current_only = F)
}
\description{
Get ICD10 parent codes (see get_parent_codes)
}
