% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{get_parent_codes}
\alias{get_parent_codes}
\title{Get parent codes from the supplied clinical dictionary}
\usage{
get_parent_codes(dict, code, immediate_parents = F, active_only = T)
}
\arguments{
\item{dict}{Clinical dictionary object}

\item{code}{Clinical code to get parent codes for}

\item{immediate_parents}{T/F flag to return only parents from only the level above (default FALSE)}

\item{active_only}{T/F flag to return only active codes (default TRUE)}
}
\description{
Get parent codes from the supplied clinical dictionary
}
\examples{
\dontrun{
config_file<-paste0(system.file("clinconcept"),"/extdata/dictconfig.json")
dict<-cc_from_file("NHSReadV3",config_file)
h3_parent_codes<-get_parent_codes(dict,"H3...",immediate_parents=F)

}
}
