% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{get_ptable_name}
\alias{get_ptable_name}
\title{Get dictionary-specific parent table name}
\usage{
get_ptable_name(dict)
}
\arguments{
\item{dict}{Clinical dictionary object}
}
\value{
table name
}
\description{
Get dictionary-specific parent table name
}
