% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{get_ptable_parent_field}
\alias{get_ptable_parent_field}
\title{Get dictionary-specific parent table parent code field name}
\usage{
get_ptable_parent_field(dict)
}
\arguments{
\item{dict}{Clinical dictionary object}
}
\value{
parent code field name
}
\description{
Get dictionary-specific parent table parent code field name
}
