% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{get_relation_codes}
\alias{get_relation_codes}
\title{Internal recursive function used to get all relation codes from the database (called explicitly by get_child_codes)}
\usage{
get_relation_codes(src, rel_code, immediate_relations = F, children = T,
  tbl_name, code_field, parent_code_field, code_data_obj)
}
\arguments{
\item{rel_code}{Code to obtain relations for}

\item{immediate_relations}{indicates if only one level (e.g. just immediate children) should be returned}

\item{config}{rexceed config object (see new_config)}
}
\value{
Character vector containing relation codes (this excludes the code provided to the function)
}
\description{
Internal recursive function used to get all relation codes from the database (called explicitly by get_child_codes)
}
