% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{get_relationships}
\alias{get_relationships}
\title{Get child/parent pairs in a data.frame for a specific code}
\usage{
get_relationships(dict, code, children)
}
\arguments{
\item{dict}{Clinical dictionary object}

\item{code}{Clinical code to check}

\item{children}{T/F flag to search child codes (T) or parent codes (F)}
}
\value{
data.frame of child/parent codes
}
\description{
Get child/parent pairs in a data.frame for a specific code
}
\examples{
\dontrun{
dict<-cc_from_file("/path/to/dictconfig")
h3_rels<-get_relationships(dict,"H31..",children=F)
}
}
