% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{load_sql_statements_from_file}
\alias{load_sql_statements_from_file}
\title{Utility function which reads SQL statements from a file and does basic placeholder replacement of
<placeholder> type strings in the statements.}
\usage{
load_sql_statements_from_file(filename, replacements)
}
\arguments{
\item{filename}{SQL file to load and process}

\item{replacements}{List of key-value replacements using the placeholder names as key.}
}
\description{
Utility function which reads SQL statements from a file and does basic placeholder replacement of
<placeholder> type strings in the statements.
}
