% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_concepts}
\alias{search_concepts}
\title{Search clinical codes in the specified dictionary using dplyr filter options}
\usage{
search_concepts(dict, ..., include_synonyms = F, output = "tbl")
}
\arguments{
\item{dict}{dictionary object (see cc_from_list/cc_from_file)}

\item{...}{Query terms passed to dplyr::filter function}

\item{include_synonyms}{Flag to include synonyms in filtering}

\item{output}{Output type. tbl is the default. Can also return codes or terms}
}
\value{
Depends on output parameter. Default is a dplyr::tbl which can undergo additional processing. Alternatives include a vector of terms or codes.
}
\description{
Search clinical codes in the specified dictionary using dplyr filter options
}
\examples{
\dontrun{
config<-cc_from_file("/path/to/dictconfig")

#return all read codes containing 'asthma' and 'lung'

 asthma_lung_codes<-search_concepts(config,term \%like\% "asthma" | term \%like\% "lung",include_synonyms=F)
}
}
