% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{write_table_from_file}
\alias{write_table_from_file}
\title{Wrapper function which simplifies reading a delimited file and writing it to a database table. Generally used with SQLite.}
\usage{
write_table_from_file(src, table_name, filename, col_names = TRUE,
  col_types = NULL, delim = ",", file_col_types = NULL, header = F,
  skip_rows = 0)
}
\arguments{
\item{src}{Database dplyr::src/connection}

\item{table_name}{Name of table in database}

\item{filename}{Filename of delimited file}

\item{col_names}{Column names passed directly to readr::read_delim}

\item{col_types}{Column types passed directly to DBI::dbWriteTable}

\item{delim}{Field delimiter to use to read file passed directly to readr::read_delim}

\item{file_col_types}{File column types passed directly to readr::read_delim}

\item{header}{Column header parameter passed directly to dbWriteTable}

\item{skip_rows}{Number of rows to skip (default = 0)}
}
\description{
Wrapper function which simplifies reading a delimited file and writing it to a database table. Generally used with SQLite.
}
