% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{test_mysql}
\alias{test_mysql}
\title{Utility function which switches on/off mysql testing}
\usage{
test_mysql(value = NULL)
}
\arguments{
\item{value}{Boolean (T/F or TRUE/FALSE)}
}
\value{
TRUE if debugging is on
}
\description{
Utility function which switches on/off mysql testing
}
\examples{

#switch debugging on
cc_debug(TRUE)

#log something if debugging is on
if(cc_debug()) {
  print("debugging is on")
}

}
